Update README.md
we are agroup of silkroad users and are trying to rebuild the site for the use of thousands but we are in need of help 
we need a full ecommence php script icluding escrow, accounts, feedback and resoulution centre. we have a hole community 
of the orginal silkroad backing us and awaiting a new site we do this in the name of Dread Pirate Roberts who gave his
freedom for this site and we want to show them we will keep fighting the battle till we win but need your help to write
the site we have the server ready and just await the software 

if you wish to help you can start right here. thank you for your time. we will stand strong and most importantly together
at this sad and hard time